Fix type of id from int to long

+++ API/Interfaces/IGamesService.cs
@@ -6,7 +6,7 @@ namespace API.Interfaces
 {
     public interface IGamesService
     {
-        Task<VideoGame> GetGameAsync(int id);
+        Task<VideoGame> GetGameAsync(long id);
         string GetImageUrl(string imageId, GameCoverSize size);
         Task<IEnumerable<VideoGame>> SearchGamesAsync(string searchString);
     }
